Add entrypoint script
